comment error handling commit

+++ services/comments.service.js
@@ -4,6 +4,8 @@ class CommentsService {
   commentsRepository = new CommentsRepository();
 
   createComment = async (comment, user_id, postId) => {
+    const existsPost = await this.commentsRepository.findPostById(postId);
+    if (!existsPost) throw new Error("404/게시글이 존재하지 않습니다.");
     const createCommentDb = await this.commentsRepository.createCommentDb(
       comment,
       user_id,
@@ -13,6 +15,8 @@ class CommentsService {
   };
 
   findComments = async (postId) => {
+    const existsPost = await this.commentsRepository.findPostById(postId);
+    if (!existsPost) throw new Error("404/게시글이 존재하지 않습니다.");
     const allComments = await this.commentsRepository.findComments(postId);
 
     allComments.sort((a, b) => {
@@ -24,6 +28,7 @@ class CommentsService {
         commentId: comment.comment_id,
         userId: comment.UserId,
         nickname: comment.user.nickname,
+        comment: comment.comment,
         createdAt: comment.createdAt,
         updatedAt: comment.updatedAt,
       };
@@ -31,6 +36,18 @@ class CommentsService {
   };
 
   putComment = async (postId, commentId, user_id, comment) => {
+    const existsPost = await this.commentsRepository.findPostById(postId);
+    if (!existsPost) throw new Error("404/게시글이 존재하지 않습니다.");
+    const existsComment = await this.commentsRepository.findCommentById(
+      commentId
+    );
+    if (!existsComment) throw new Error("404/댓글이 존재하지 않습니다.");
+    const existsCommentById = await this.commentsRepository.findCommentByBothId(
+      commentId,
+      user_id
+    );
+    if (!existsCommentById)
+      throw new Error("403/댓글의 수정 권한이 존재하지 않습니다.");
     const putComments = await this.commentsRepository.putCommentDb(
       postId,
       commentId,
@@ -41,6 +58,20 @@ class CommentsService {
   };
 
   deleteComment = async (postId, commentId, user_id) => {
+    const existsPost = await this.commentsRepository.findPostById(postId);
+    if (!existsPost) throw new Error("404/게시글이 존재하지 않습니다.");
+    const existsCommentById = await this.commentsRepository.findCommentByBothId(
+      commentId,
+      user_id
+    );
+    const existsComment = await this.commentsRepository.findCommentById(
+      commentId
+    );
+    if (!existsComment) throw new Error("404/댓글이 존재하지 않습니다.");
+
+    if (!existsCommentById)
+      throw new Error("403/댓글의 삭제 권한이 존재하지 않습니다.");
+
     const deletedComment = await this.commentsRepository.deleteCommentDb(
       postId,
       commentId,
